Rename directory to be a valid python package
