+++ ruggedpod_api/common/exception.py
