+++ setup.cfg
@@ -10,8 +10,8 @@ classifier =
 
 [files]
 packages =
-    ruggedpod-api
+    ruggedpod_api
 
 [pytest]
-addopts = ruggedpod-api --pep8
+addopts = ruggedpod_api --pep8
 pep8maxlinelength = 120
